+++ params/network_upgrades.go
@@ -194,7 +194,7 @@ func (n NetworkUpgrades) Description() string {
 	return banner
 }
 
-func getNetworkUpgrades(agoUpgrade upgrade.Config) NetworkUpgrades {
+func GetNetworkUpgrades(agoUpgrade upgrade.Config) NetworkUpgrades {
 	return NetworkUpgrades{
 		ApricotPhase1BlockTimestamp:     utils.TimeToNewUint64(agoUpgrade.ApricotPhase1Time),
 		ApricotPhase2BlockTimestamp:     utils.TimeToNewUint64(agoUpgrade.ApricotPhase2Time),
